Add gathering, mining, and session listeners for custom statistics tracking
- Implement GatheringListener to track collected berries and honey.
- Implement MiningListener to track block breaks, rare blocks, and harvested plants.
- Implement SessionListener to maintain login streak statistics.
- Introduce StatContext, StatDefinition, StatKind, and StatRegistry for managing statistics.
- Create storage classes for JDBC (MySQL and SQLite) and YAML.
- Implement UpdateChecker for asynchronous update checks and notifications.
- Add UpdateNotifyListener to inform OPs about available updates upon joining.
- Include .gitignore for build artifacts and IDE files.

+++ playerstats/src/main/java/it/markus/playerstats/stat/StatResult.java
@@ -1,7 +1,10 @@
 package it.markus.playerstats.stat;
 
 /**
- * Ein Statistik-Ergebnis fuer einen Spieler (Name + Rohwert).
+ * Ein Statistik-Ergebnis fuer einen Spieler (Name + Wert).
+ *
+ * Der Wert ist {@code double}, damit auch berechnete Statistiken (K/D,
+ * Trefferquote) ohne Sonderbehandlung sortier- und darstellbar sind.
  */
-public record StatResult(String playerName, long value) {
+public record StatResult(String playerName, double value) {
 }
